refactor: split JS into components
