+++ src/tokens.ts
@@ -1,0 +1,4 @@
+export namespace CommandIDs {
+  export const restartRunStateless = 'gennaker-tools:restart-run-stateless';
+  export const resetJupyterLab = 'gennaker-tools:reset-jupyterlab';
+}
